+++ expectedDistribution.csv
@@ -1,0 +1,6 @@
+projet,type,Archived,adm1,adm2,adm3,adm4,Name,Date of distribution ,Update on ,Selection criteria,Commodities ,Distribution beneficiaries
+Dev Project,0,1,,,,,maison,2018-08-01 00:00:00,2018-08-01 08:13:58,,,
+Prod Project,0,1,,,,,nourriture ,2018-08-01 00:00:00,2018-08-01 08:13:58,,,
+Dev Project,0,1,,,,,maison,2018-08-01 00:00:00,2018-08-01 08:13:58,,,
+Prod Project,0,,,,,,salade,2018-08-01 00:00:00,2018-08-01 08:13:58,,,
+Prod Project,0,,,,,,salade,2018-08-01 00:00:00,2018-08-01 08:13:58,,,
